[AIRLFOW-YYY] Move db commands to separate file
