+++ airflow/cli/commands/db_command.py
@@ -1,0 +1,44 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+"""Database sub-commands"""
+from airflow import settings
+from airflow.utils import cli as cli_utils, db
+
+
+def initdb(args):
+    """Initializes the metadata database"""
+    print("DB: " + repr(settings.engine.url))
+    db.initdb()
+    print("Done.")
+
+
+def resetdb(args):
+    """Resets the metadata database"""
+    print("DB: " + repr(settings.engine.url))
+    if args.yes or input("This will drop existing tables "
+                         "if they exist. Proceed? "
+                         "(y/n)").upper() == "Y":
+        db.resetdb()
+    else:
+        print("Bail.")
+
+
+@cli_utils.action_logging
+def upgradedb(args):
+    """Upgrades the metadata database"""
+    print("DB: " + repr(settings.engine.url))
+    db.upgradedb()
