+++ tests/cli/commands/test_db_command.py
@@ -1,0 +1,40 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+import unittest
+from unittest import mock
+
+from airflow.bin import cli
+from airflow.cli.commands import db_command
+
+
+class TestCliDb(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.parser = cli.CLIFactory.get_parser()
+
+    @mock.patch("airflow.cli.commands.db_command.db.initdb")
+    def test_cli_initdb(self, initdb_mock):
+        db_command.initdb(self.parser.parse_args(['db', 'init']))
+
+        initdb_mock.assert_called_once_with()
+
+    @mock.patch("airflow.cli.commands.db_command.db.resetdb")
+    def test_cli_resetdb(self, resetdb_mock):
+        db_command.resetdb(self.parser.parse_args(['db', 'reset', '--yes']))
+
+        resetdb_mock.assert_called_once_with()
